usize instead of u32

+++ src/parser.rs
@@ -12,7 +12,7 @@ pub enum SemType {
     Subexpression,
     LitTerm,
     Literal(String),
-    Term(u32)
+    Term(usize)
 }
 
 /*
